requirements.txt
fastapi==0.103.1
uvicorn==0.22.0
motor==3.1.1
pymongo==4.3.3
pydantic[email]==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-dotenv==1.0.0
